refactor(cef): carve window motion handlers into window/motion.rs
Second sub-PR of the commands/window.rs modularization (Plan 1). Pure
move — no behavior change.

Extracted to window/motion.rs: get_window_position, move_window_by,
set_window_position, start_window_drag, resolve_window_at_cursor,
update_floating_redock_hover, clear_floating_redock_hover. All `pub`,
dispatched by ipc.rs (re-exported `pub use motion::*`).

These were a contiguous block, so the carve is a clean cut. motion.rs
imports the two HWND helpers it needs from the sibling `lifecycle`
module (`resolve_window_hwnd` pub(super), `find_own_top_level_window`
pub(crate)) under a `#[cfg(windows)]` use. mod.rs dropped
`resolve_window_hwnd` from its lifecycle import (only the motion fns
used it); `find_all_own_windows` stays (transparency handlers still
in mod.rs).

Verified: `cargo check -p agentmux-cef --release` clean.
