+++ .changesets/1780032994-refactor-cef-extract-window-motion-handlers-into-commands-wi-hr45.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(cef): extract window motion handlers into commands/window/motion.rs
